requirements.txt
streamlit
tensorflow
matplotlib
scikit-learn
pillow
numpy
